FEATURE: rename MovieBookingPage.tsx to MovieBooking.tsx, and update main.tsx based on rename
